Added restaurant bill calculator and alarm calculator
There are 2 programs. The first will ask the user to enter the meal total and will calculate a new total that will include tip and tax for the meal. The second will ask a user to input the current time in hours and future time for an alarm based on 24 hour clock . The program will output the time in hours the alarm will sound based on a 24 hour clock.